+++ wallet/src/main/java/com/bbva/wallet/services/AuthenticationService.java
@@ -8,14 +8,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 
-import com.bbva.wallet.entities.Account;
 import com.bbva.wallet.entities.Role;
 
 import com.bbva.wallet.enums.Currency;
 import com.bbva.wallet.enums.RoleName;
-import com.bbva.wallet.repositories.AccountRepository;
 import com.bbva.wallet.repositories.RoleRepository;
-import com.bbva.wallet.utils.Utils;
 
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -26,11 +23,9 @@ public class AuthenticationService {
     private final AuthenticationManager authenticationManager;
 
     private final UserRepository userRepository;
-    private final AccountRepository accountRepository;
+    private final AccountService accountService;
     private final PasswordEncoder passwordEncoder;
     private final RoleRepository roleRepository;
-    private final Utils utils;
-
 
     public User signUp(UserSignUpDTO userSignUpDto) {
         //Role role = new Role(RoleName.USER);
@@ -53,27 +48,11 @@ public class AuthenticationService {
                 .build();
 
         userRepository.save(user);
-        Account accountPesos = Account.builder()
-                .currency(Currency.ARS)
-                .transactionLimit(300_000.0)
-                .balance(0.0)
-                .user(user)
-                .cbu(utils.generateRandomCbu())
-                .build();
-        Account accountDolares = Account.builder()
-                .currency(Currency.USD)
-                .transactionLimit(1_000.0)
-                .balance(0.0)
-                .user(user)
-                .cbu(utils.generateRandomCbu())
-                .build();
 
-        accountRepository.save(accountPesos);
-        accountRepository.save(accountDolares);
+        accountService.createAccount(Currency.ARS, user);
+        accountService.createAccount(Currency.USD, user);
 
         return user;
-
-
     }
 
     public User logIn(UserLogInDTO userLogInDTO) {
